python/requirements.txt
mysql-connector-python>=8.0.0
pytest>=6.0.0
pytest-mock>=3.0.0
